Couple of ideas to further improve the builder
I really like what @ginkgomzd with the build process - I'm not going to
lie, this is my first Makefile, and I appreciate the cleanup.  I had a
few ideas looking it over:

If we're going to build a package during the install process,
why not install the package that we just built?  I think that works
better anyway - once you test the package that was installed, you can
confidently distribute it knowing that it is the same code you just ran.

I also  added a clean step to delete all of the generated files
(linodecli/data-* and linode-cli.sh along with the contents of dist/) to
make sure everything's fresh with each build.