Dynamically set default Linode type for deployed K8s masters (resolves #93)
Currently, we set hardcoded instance types for both nodes and
masters. However, the linode-cli lets you specify a default type for
newly created Linodes.

Ideally, we would like to respect that choice, however K8s masters do
require 2 VCPUs at minimum. Since the available plans are available
through the API, we can look up the user's configured default type to
see if it meets this constraint. If it does, it's fine to use that. If
it doesn't, we pick the smallest plan that meets this constraint. (In
this case the user's default is smaller than the minimum, so they're
probably looking to deploy smaller Linodes if possible.)