Added support for plugin configuration
Plugins may want to store persistent configuration data on a per-user
bases.  Instead of having them manage that themselves, this allows
plugins to use the Linode CLI's config in a uniform manner to store
persistent data.

All plugin config items' keys begin with `plugin-` followed by the
plugin's name.  Keys are stored per-user in the config, so multiple
users may have different settings configured.

This is not presently in use, but it might come in handy for some
upcoming plugins that may or may not be in the works ;)