Added third-party plugin support
This adds a new command, `register-plugin`, which loads the python
module supplied by the user and checks that it is a runnable CLI plugin
before registering it to the CLI's config.  On further invocations of
the config, this plugin will be available.

Third party plugins require two things:

1. A `call` function that matches the expected signature
1. A `PLUGIN_NAME` constant that defines how the plugin should be called

This PR needs the following:

- [ ] Documentation
- [ ] Handling naming conflicts between plugins
- [ ] Re-registering