Adds `ssh` plugin
This replaces #72.

Adds a new plugin, `ssh`, which is invoked as follows:

```bash
linode-cli ssh root@my-linode-label
```

The username piece is optional and defaults to the user invoking the CLI
(as normal ssh behavior).  This is basically a proxy for finding a
public IP address to ssh into the Linode with by search for a Linode
with the given label.

If no Linode with the given label exists but other Linodes are found
that are close, those labels are output as a suggestion.

All arguments after the `[USERNAME@]LINODE_LABEL` are passed on to ssh
as normal.

Big thanks to @displague and @jfrederickson for their feedback regarding
this plugin.