Only suggest an upgrade when an upgrade exists
Closes #130

Right now, when a new version of the Linode API ships, there is some
span of time where the CLI knows it's out of date, but a new version
isn't available.  After some discussion and thought, this change makes
the CLI check PyPI for a newer version before warning that it should be
upgraded.  This check is only performed when the API response's headers
indicate that a new spec version is available.

Also refactored so that all version checking is disabled if
`--suppress-warnings` is used, since it won't matter in that case.

This adds a single additional HTTP request to CLI calls when the API
indicates a new spec version, which is _probably_ okay.  I'm open to
opinions on this.

I tested this by changing the version the CLI build thought it was, and
the spec version it was built off of.  This can't _really_ be tested
until the next API release.