Check Terraform Version in K8s Plugin
This closes #131

After the Terraform command as been found, we can run `terraform
version` and compare it against our minimum and maximum versions.
Currently, those versions are v0.11.14 and v0.12.0.

The minimum version is an inclusive check, where v0.11.14 is supported.
The maximum version is an exclusive check, where v0.12.0 is not
supported.

It is not clear if Terraform uses true semantic versioning, but we
assume that the "major.minor.patch" are all going to be integers for now.