Edit Makefile to fix issue #78 and allow py2/py3
As described in Issue #78, the existing makefile expects "python" to
point to python2. However, on many modern distros "python" actually
points to python3 (and python2 is available by "python2"). This changes
the python2 command to "python2" if the python version detected is
python3. (This *could* create a different issue if python2 isn't
    available with "python2" ... but I *believe* that it should be in
    99-100% of cases?)