Don't require a config file when using an environment-defined token
Closes #140

Due to an oversight, POST and PUT operations would attempt to pull
defaults form the config file even if the CLI was never configured and
we running with the `LINODE_CLI_TOKEN` environment variable defining the
token for it to use.  This could break some automated setups that never
had an interactive tty without using the `--no-defaults` flag, but that
shouldn't be required.

This change tracks if the CLI used an environment-defined token and
skips updating the namespace sent to the operation if the config wasn't
loaded in that case.