bug: Fix filtering by tags (and other arrays)
Closes #151

This was introduced in #141 - by parsing filter arguments as their
declared types in the spec, we were unintentionally turning "tags" into
an array of characters (it's defined as an array in the spec), so
`--tags test` would result in an header of `X-Filter: '{"tags":["t","e","s","t"]}'`,
which the API would balk at (correctly).

This change now captures the declared _item type_, and expects that for
any filterable attribute that is declared as an array.

@patthiel, we should add a test for filtering requests with tags to
prevent future regressions in the same vein.