Handle lists of objects
LKE uses node_pools, a list of objects, to specify what pools must be
supplied when creating an LKE Cluster.  This is the first time we've had
such input, and the CLI didn't handle it well.

This is a slightly scary change as it changes how the spec is parsed -
please test carefully with other endpoints!