Improved JSON output
This closes #153

Presently, `--json` simply outputs the normal fields as a JSON object,
with no special formatting or considerations, and it's up to the caller
to request `--pretty` and/or `--all` depending on their use-case.  This
change now enables both `--pretty` and `--all` for all JSON output.

`--all` can still be overridden via `--format`, which, if provided, will
trim down the resulting JSON object to only the named fields.

`--pretty` can not be disabled in this initial commit.  I don't _think_
that's breaking, as it still outputs valid JSON (I can still `| jq`, for
instance), but I'd like some more eyes on it as I don't want to break
scripted usages with this change.