Handle multipart uploads correctly
Closes #174

Previously, large files were sent as normal uploads, causing a long
upload time followed by a `EntityTooLarge` error.  This is because s3
has a maximum size for file uploads; this isn't the maximum size for
objects, mind, but for individual uploads.  Files larger than this size
must be uploaded in multiple parts.

This change supports uploading files as multiple parts.  I tested it
with some fabricated large files made with this command:

```bash
dd if=/dev/zero of=file-to-upload bs=1024 count=0 seek=20240
```