Allow API POST/PUT Body Elements to Conflict with CLI Flags
Closes #154

For the first time ever, one of the API's request arguments conflicts
with a built-in CLI flag, that being `POST lke/clusters`'s argument
"version".  When translated to the CLI's interface, it becomes
`--version`, which the CLI will pick up as a request to print its own
version number an exit.

This change adds a new spec extension to Linode's OpenAPI spec,
`x-linode-cli-name`, which is allowed in schemas and defines the name
the CLI should use for that element _instead of_ the name of that schema
in the spec.  This should be used only in exceptional cases (like that
noted above) to keep the CLI's behavior predictable.