Allow regenerating Object Storage keys
Closes #178

Previously, there was no way to regenerate your object storage keys
without manually editing the CLI's config, which is a pretty lousy user
experience.  If you were to delete the keys in cloud, for example, the
CLI would be stuck attempting to use them for requests to Object
Storage.

This change adds the "regenerate-keys" command to the `obj` plugin,
allowing object storage keys to be regenerated in case they become
expired or otherwise don't work.