Make total size human-readable in `linode-cli obj du` output
Closes #181

When the object storage plugin lists bucket usage via the `du` command,
the size of each bucket is converted in KB, MB, or GB to present a more
human-friendly interface.  Due to an oversight, the total size of all
buckets was not converted to a human-friendly value in the same way, and
is being displayed simply as total bytes.

This change refactors the code that did the conversion out into its own
function, and calls it for both bucket totals and the grand total of
usage in a cluster.