bug: Handle paginated responses without $refs
The CLI's spec parser currently makes a lot of assumptions, and one is
that paginated responses will have a `$ref` for their data, instead of
listing the schema directly.  This isn't required by OpenAPI, and for
the first time it isn't the case in our spec.

This change handles non-`$ref` paginated response data.