Accept passwords from environment
Closes #200

As it is now, the CLI either accepts passwords on the command line, or
interactively.  The interactive option was added to allow more secure
password entry, but it doesn't cover the automated case as mentioned in
the above issue.

This change implements the suggestion included in #200 to accept
passwords via environment variables to allow secure entry in a
non-interactive setting.  The CLI now gets the values for password
fields like so:

* If a password was given on the command line, use it
* If not, is an environment variable is set for the field, use its value
* If not, prompt for a password

This preserves backwards-compatible behavior while allowing the
environment to be considered.

The CLI will expect environment variables to match the CLI's argument
name, all uppercase, prefixed with `LINODE_CLI_` - for example, the
`--root_pass` argument to `linodes create` will look for
`LINODE_CLI_ROOT_PASS` in the environment.