Handle Prefixes in Obj Plugin
Previously, the obj plugin skipped objects with `/` in their names,
which is a somewhat perplexing choice I'd made when originally
implementing it.  This change brings the behavior more inline with s3cmd -
objects are listed with `/` as a delimiter, and prefixes can be
specified after the bucket name.  For instance:

```
$ linode-cli obj ls prefix-test
                  DIR  prefix/
2020-08-11 17:50  8    testfile3
$ linode-cli obj ls prefix-test/prefix/
2020-08-11 17:50  4  prefix/testfile
2020-08-11 17:50  5  prefix/testfile2
```