new: Accept StackScript scripts from file
Closes #201

Previously, only "ssl-cert" and "ssl-key" formatted fields were
accepted as file paths (and then only if the input looked like a path
and a file existed at it); this change brings that behavior over to
other fields.  The linked docs PR below ties this into the StackScript
"script" field, but it can be used with any field now using a new spec
extension: `x-linode-cli-format`.  As per the included documentation in
this PR, the only valid value is "file" at present; the intention is
that this overrides the "format" given in the OpenAPI spec, but only in
the CLI, and only if present.  Any field with this format will inherit
the above-described path-checking behavior.