Don't send empty lists in requests
This was pointed out by @adammohammed

This impacts the `object-storage keys-create` operation, and by
extension the `obj` plugin for new users.

This behavior didn't matter before now, which is probably why it went
unnoticed.  The new `bucket_access` argument to `POST /object-storage/keys`
does behave differently when given an empty list, however, and sending
it along caused the generated keys to have no access to anything.