Run --help without a config in CLI and obj plugin
Closes #214

Previously, the CLI Configuration was initialized expecting to load a
config even if `--help` was passed in, which is unintuitive behavior
because the config will prompt for configuration immediately if a file
isn't found.  This was further complicated by the `obj` plugin, which
manages its own configuration and would fail if the CLI wasn't
configured when it was invoked, even with only a `--help` command.

This change is intended to make unconfigured user of the CLI easier by:

 - Allowing `--help` commands to run without configuration
 - Allowing the `obj` plugin to run `--help` commands without
   configuration
 - Providing better errors in the `obj` plugin, namely around cluster,
   to allow users without a config to easily identify and remedy common
   issues.

This should be tested thoroughly, as it touches some fundamental parts
of the CLI's behavior; namely, you should verify that:

 - Configured use works as it does today
 - Calling with `--skip-config` and `--no-defaults` behaves as expected
 - Using environment variables for tokens and obj keys works as expected
 - Incorrectly setup environments for the above produce useful errors