Move to openapi3 for Linode CLI
This is for us.

This requires:

- [ ] https://github.com/Dorthu/openapi3/pull/21
- [ ] https://github.com/Dorthu/openapi3/pull/24
- [ ] https://github.com/Dorthu/openapi3/pull/25

Currently, the Linode CLI uses an OpenAPI parser that was written to our
OpenAPI spec (at the time the CLI was written) and evolved slowly as
issues were found.  What's resulted is a very organic, free-range parser
that exists specifically for the Linode CLI and is rather difficult to
unwind and maintain.

This change removes the CLI's spec parser in favor of the [openapi3
python package](https://github.com/Dorthu/openapi3), which I wrote after
completing the CLI as a proper spec parser, in part to replace the one
in this project.  The openapi3 package [is already used to validate
Linode's OpenAPI specification during automated builds](https://github.com/linode/linode-api-docs/blob/9da259cbf7f50c3d26fec1f9a227337c6dc87068/.travis.yml#L6),
meaning that transitioning to this parser will ensure the doc's builds
never fall out of sync with the CLI's capabilities.

:warning: This is a major change, and will require manual testing.

Additional considerations:

- [ ] Ensure all CLI-specific spec extensions are still supported
- [ ] Ensure all first-party plugins function as needed
- [ ] Adjust requirements for installation as necessary
- [ ] How large of a version bump should this be? (in theory it's only a
      refactor, but it also has a lot of potential to break things
      unintentionally).