new: Added image-upload first-party plugin
The new POST /images/upload endpoint allows users to initiate an image
upload and returns a URL to upload the image data to.  While this can be
worked out on the command line, it's much more convenient to have a
plugin that does the entire process, accepting a file path, calling
the new endpoint, and actually handling the upload all in one command.

This new endpoint must be added to the docs before it can be supported
in this plugin.  During the beta period, this endpoint will not be
available to all users, so this plugin offers a reasonable error message
for users who attempt to invoke it and do not yet have access to it.