Don't require configuration when running `--version`
Closes #235

While `--skip-config` does allow you to find the version without a token
provided, there's no reason to prompt for it at all, and I can see how
this would be disruptive.

This change treats `--version` as `--skip-config` when deciding if the
user should be prompted for configuration if none is present.