bug: Fixed populating defaults when no configuration is present
Closes #239

If using the CLI with an environment-defined token (`LINODE_CLI_TOKEN`),
populating defaults in a request would fail with a
`configparser.NoSectionError` as it attempted to load defaults for an
unconfigured user.

This change checks to see if the user is configured before attempting to
apply defaults, and if so returns the namespace given instead of
attempting to update it.