fix: Check Image upload size before attempting upload
Closes #250

Right now, the `image-upload` plugin just attempts to upload whatever
it's given, but that can waste a lot of time.

This change checks that the compressed size is no larger than the
maximum allowed (5GB) before initiating the upload.

This change does not attempt to catch cases where the _uncompressed_
size of the image is larger than allowed; the maximum uncompressed size
is not constant, and the CLI has no way to know if an Image uploaded
will end up being too large when it gets there.