Remove dependency on colorclass
Closes https://github.com/linode/linode-cli/issues/231

The colorclass module is not compatible with Python 3.10, and is
currently [archived on github](https://github.com/Robpol86/colorclass)
with issues open about future compatibility issues and no real
suggestion of movement.

This change replaces `colorclass` with the few color codes that were
actually used so far, which removes the dependency entirely and should
allow the CLI to work on python 3.10.

This appears to work where I've tested it, and from what I read it aught
to work the same on Windows too, but I'd like someone to give it a spin
first to be sure.