bug: Resolve nested refs
Closes #255

The spec for LKE Node Pools included deeper `$ref` nesting than the
CLI's parser supported.  This change resolves one more level of `$ref`,
allowing node pools counts to be updated easily.

This isn't an awesome fix; the real problem here is the fragility of the
CLI's spec parser, and the real fix is to implement a better spec
parser, as #218 attempted to do.  That effort should be completed to
resolve all like issues that may be found in the future.