bug: Only allow requested defaults to actions
This should fix the `linode-cli tickets create` action.

The Linode CLI, when being configured, accepts default values for
"region", "image", and "type".  These values were previously added to
every input for a PUT or POST endpoint, and then parsed out as normal.
This had the effect of defaulting those values to the configured
defaults for actions like `linodes create` or `nodebalancers create`
without additional interaction.

However, the `tickets create` action recently began accepting
`--region`, but only for certain types of tickets.  This caused ticket
creation to fail through the CLI if a default region was configured,
unless `--no-defaults` was given, or unless the specific ticket type was
being opened.

This change allows the spec to specify which actions should use which
defaults, allowing this problem to be resolved as well as avoiding it
for future changes to actions.

:warning: This requires a docs update to allow existing actions to
continue using their defaults.  Such a PR will be linked below.