bug: Fix parsing of nested objects in lists
This fixes tests, and the lke cluster-create operation.

As it happens, the CLI did not parse optional list arguments correctly
(namely, if an optional argument to a list was omitted, the list would
not be created and no items in the list would populate correctly.

Compounding this, objects nested in objects within lists did not
properly restructure after being flattened for the command line.  This
caused nested objects in lists to be sent incorrectly.

Both of the above broke the `linode-cli lke cluster-create` endpoint,
and thereby tests, when `autoscaler` was added to that endpoint, nested
within the `node_pools` object in a list.  This change fixes the issues,
allowing this endpoint to work again.

:warning: I did not write this to handle multiple layers of object
nesting in lists.  I will revisit this if it ever becomes necessary - I
left a comment noting where it would be handled.