More progress on proper spec parsing
Abandoned the concept of storing the parsed OpenAPI object, as pickling
it didn't work as intended.  Working on restructuring code so that baked
data objects are separate from runtime logic to keep the distinction
between what is safe to pickle and what isn't easier, and just to
improve structure overall.