bug: Tests should accept new error output
This is related to https://github.com/linode/linode-cli/pull/274

In #274, I changed the output when a command is entered improperly to
include the command and action the CLI was processing when the error
occurred, which mimics what was entered on the terminal.  I didn't
realize tests were checking for the old output.

This change updates tests to look for the new output instead of the old
format.