bug: Fix `linode-cli users view USERNAME`
In the process of attempting to free the namespace for CLI Args that
conflict with URL Params, I accidentally overwrote the stored param
string for all methods under a given path.  This caused the operation
`linode-cli users view` to have a param `username_`, but the URL to
continue to be `/account/users/{username}`, as the param's name was
updated when processing `linode-cli users update` (which has a
`--username` arg and a `username` param, causing the param to change to
`username_` and the url to change to `/account/users/{username_}`).

This change copies all params, so that when they are updated, only the
copy local to this method is updated.