new: Warn python2 users of planned dropping of support
Python 2 was sunset on Jan 1st, 2020.  We have continued to support
python2 this long because many users still use it, however as python3
support has grown and libraries continue dropping support for python2,
it is becoming increasingly difficult to maintain the CLI with python2
supported.

This change adds a warning to users running the CLI under python2 that
support will be dropped in a future version, and includes a link to our
official guide on how to upgrade your installation.  As usual,
`--suppress-warnings` silences this warning, and the warning is output
to `stderr` to allow users to redirect `2>/dev/null` to ignore it as
well.

In the last version that supports python2, this warning will be changed
to note that support _has been_ dropped, and that upgrading to a python3
install is required to continue receiving updates.

I implemented this to print the warning on every request, however I also
considered having this warning only display on the first request of a
given day.  I'm open to opinions on this.

This additional output could break scripted installations, for example
those doing something like `linode-cli linodes list --text --format
'id,ipv4 | something` if those installations do not already ignore
stderr/suppress warnings.  I find this to be unlikely, as the CLI
already prints warnings when an update is available, so anything that's
been running unmonitored for any amount of time would already break if
warning output wasn't being handled.  Again, opinions are welcome.