new: Add support for `/acctest` command (#397)
This change adds support for the `/acctest` command, which allows
maintainers to run integration tests on PRs using a command and commit
hash. This implementation mimics the implementation for other DX-owned
repos (terraform-provider-linode, ansible_linode, etc.)

The following environment variables will need to be added to the secrets
of this repository:
- DX_ACCTEST_APP_ID
- DX_ACCTEST_PRIV_KEY

**NOTE: This change excludes bats tests due to the planned removal of
bats in the near future. See #394**