Replace terminaltables with rich (#440)
## 📝 Description

Relates to #436 

The version of `terminaltables` used by this project has been archived
for a few years now and was recently removed from the Arch community
repository (which is what led me to embark on this PR). While the
project may remain viable for some time to come, I propose the
maintainers of `linode-cli` consider switching to
[rich](https://github.com/textualize/rich), which offers much greater
versatility when rendering rich text to the terminal in Python (even
`pip` uses rich for its progress bar functionality).

I have made the necessary changes to the code in this PR.

### Summary of changes:
- Updated `requirements.txt`
  - Added `rich`
  - Removed `terminaltables`
- Updated all instances of `terminaltables.SingleTable` objects
throughout the project with equivalent `rich.Table` objects. The
following modules were updated under this effort:
  - `arg_helpers.py`
- `output.py` - Note for this module I also added the `rich.Text` module
to enable proper ANSI sequence rendering. Without converting the table
cells to `Text` objects, any cell containing ANSI sequences will result
in improper rendering of the columns. This should not be applicable to
any other module, but could be implemented for uniformity's sake if
desired with no impact to functionality.
  - `__init__.py`
  - `plugins/obj/__init__.py`
  - `plugins/obj/helpers.py`
- Updated the `firewall-editor` plugin to replace
`terminaltables.PorcelainTable` object with `rich.Table`. The new
`rich.Table` is configured to be equivalent to the corresponding
`PorcelainTable` it's replacing.
- Updated the unit test `test_output.py` to use `rich.Table`

## ✔️ How to Test

### Steps
- Run any set of commands associated with generating tables. Among
these, make sure to run a command that generates a table with colored
text, such as `linode-cli linodes list` to ensure the ANSI escape
sequence'd text renders correctly.
- Run `linode-cli` to see the main help message tables
- Run `linode-cli firewall-editor <label>` to see the new firewall
editor table.
- Run the `test_output.py` unit test `python -m unittest test_output.py`

### What I tested
- Each of the 4 tests above

### What I **did not** test
- I did not run the full test suite as I didn't want to destroy my
account's configuration nor did I want to make a new account for this
purpose.


## 📷 Preview

### Before/After
- `linode-cli` before

![linode-cli-help-before](https://user-images.githubusercontent.com/7587293/232877125-1fa30359-24bf-4d97-b4a0-c4fc1abf3fe4.png)
- `linode-cli` after

![linode-cli-help-after](https://user-images.githubusercontent.com/7587293/232877477-3ae282b2-387a-4b72-a49a-9ca12b99c962.png)
- `linode-cli linodes list` before

![linode-cli-linodes-list-before](https://user-images.githubusercontent.com/7587293/232877191-e1091e87-77c1-4a0d-b4c4-a636da21d70d.png)
- `linode-cli linodes list` after

![linode-cli-linodes-list-after](https://user-images.githubusercontent.com/7587293/232877536-58b9232a-a88e-4dda-8560-29ab54c6cf2d.png)
- `linode-cli firewall-editor <label>` before

![linode-cli-firewall-editor-before](https://user-images.githubusercontent.com/7587293/232877301-31272971-4319-4bea-89fa-704b265ddf45.png)
- `linode-cli firewall-editor <label>` after

![linode-cli-firewall-editor-after](https://user-images.githubusercontent.com/7587293/232877620-b317ea77-d04c-4ed4-91e8-74400a8d6f4a.png)

## Other Notes

I don't understand exactly how the `Makefile` works for this project,
but something prevents me from acquiring the necessary baked data
(specifically `bake_version`) using `make` in my fork of the project. If
I run `make` from this project and copy the `baked_version` file to my
fork, I am able to install with `pip`. I did not make any changes to the
`Makefile` (or anything else other than what I listed above for that
matter), but I thought this would be worth noting in the PR.

I hope you seriously consider merging this PR and/or implementing `rich`
on your own. I can't speak highly enough about its performance and
options. Thank you for taking the time to consider these changes.

---------

Co-authored-by: Will Smith <Dorthu@users.noreply.github.com>
Co-authored-by: Lena Garber <114949949+lgarber-akamai@users.noreply.github.com>